check dir existance as powershell Remove-Item does not return exit 0 if not found

+++ scripts/windows/test/Setup-Env.ps1
@@ -68,7 +68,10 @@ $IoTEdgedArtifactFolder = Join-Path $E2ETestFolder "artifacts\iotedged-windows"
 $InstallationScriptPath = Join-Path $E2ETestFolder "artifacts\core-windows\scripts\windows\setup\IotEdgeSecurityDaemon.ps1"
 Invoke-Expression $InstallationScriptPath
 
-Remove-Item "C:\data\edgehub" -Recurse -Force
+If ((Test-Path "C:\data\edgehub"))
+{
+    Remove-Item "C:\data\edgehub" -Recurse -Force
+}
 
 If ($osEdition -eq "IoTUAP")    # Windows IoT Core - update iotedge
 {
